renamed both buildings page components for clarity
