+++ src/pages/AllBuildings.tsx
